refactor(test): bring RN test fixture packages in line with each other
