+++ test/react-native/features/fixtures/rn0.55/android/app/src/main/java/com/reactnative/MainActivity.java
@@ -1,4 +1,4 @@
-package com.rn055;
+package com.reactnative;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "rn055";
+        return "reactnative";
     }
 }
