+++ test/react-native/features/fixtures/rn0.55/android/app/src/main/java/com/reactnative/MainApplication.java
@@ -1,4 +1,4 @@
-package com.rn055;
+package com.reactnative;
 
 import android.app.Application;
